web-app/requirements.txt
Flask==3.0.2
Authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.1